update contact-loaders imports after move
