+++ src/integrations/contact-loaders/test-fakedata/react-component.js
@@ -1,12 +1,12 @@
 import type from "prop-types";
 import React from "react";
 import RaisedButton from "material-ui/RaisedButton";
-import GSForm from "../../components/forms/GSForm";
+import GSForm from "../../../components/forms/GSForm";
 import Form from "react-formal";
 import Subheader from "material-ui/Subheader";
 import Divider from "material-ui/Divider";
 import { ListItem, List } from "material-ui/List";
-import CampaignFormSectionHeading from "../../components/CampaignFormSectionHeading";
+import CampaignFormSectionHeading from "../../../components/CampaignFormSectionHeading";
 import CheckIcon from "material-ui/svg-icons/action/check-circle";
 import WarningIcon from "material-ui/svg-icons/alert/warning";
 import ErrorIcon from "material-ui/svg-icons/alert/error";
